fix: rename interface file & make test quick
